Some changes in the iterator type
1. Change were done on the iterator type. The `long` types were
   changed in `uint64_t`

2. The `observe` method of the calss `mpi_atm_sim` is now under
   maintenance. Strange issues were reported during the atmosphere
   observations. The TODs, in general, present some `bad data` and
   the function, during the observation operations, return a lot
   WARNINGS about an out-of-cone observations (not very often), and
   more often that not a "full index out of bounds" that it's
   not properly related to the observations but to the cells
   interpolation.

For the next future, the functions `mpi_atm_sim::observe` and
`mpi_atm_sim::interp` have to be double checked.

I want to remember that `cal` library was extracted by the TOAST
framework, and during this oparation a lot of features about the
processes distrubution were been lost. In this context, the
developers have to be used to work with these new framework and
they are filling every lack.

+++ src/libcal_mpi/src/smooth_interpolation.cpp
@@ -12,9 +12,9 @@ double cal::mpi_atm_sim::interp(double x, double y, double z, std::vector <long>
 {
     // Trilinear interpolation
 
-    long ix = (x - xstart) * xstepinv;
-    long iy = (y - ystart) * ystepinv;
-    long iz = (z - zstart) * zstepinv;
+    int64_t ix = (x - xstart) * xstepinv;
+    int64_t iy = (y - ystart) * ystepinv;
+    int64_t iz = (z - zstart) * zstepinv;
 
     double dx = (x - (xstart + (double)ix * xstep)) * xstepinv;
     double dy = (y - (ystart + (double)iy * ystep)) * ystepinv;
@@ -59,19 +59,19 @@ double cal::mpi_atm_sim::interp(double x, double y, double z, std::vector <long>
         }
 # endif // ifdef DEBUG
 
-        size_t offset = ix * xstride + iy * ystride + iz * zstride;
+        int64_t offset = ix * xstride + iy * ystride + iz * zstride;
 
-        size_t ifull000 = offset;
-        size_t ifull001 = offset + zstride;
-        size_t ifull010 = offset + ystride;
-        size_t ifull011 = ifull010 + zstride;
-        size_t ifull100 = offset + xstride;
-        size_t ifull101 = ifull100 + zstride;
-        size_t ifull110 = ifull100 + ystride;
-        size_t ifull111 = ifull110 + zstride;
+        int64_t ifull000 = offset;
+        int64_t ifull001 = offset + zstride;
+        int64_t ifull010 = offset + ystride;
+        int64_t ifull011 = ifull010 + zstride;
+        int64_t ifull100 = offset + xstride;
+        int64_t ifull101 = ifull100 + zstride;
+        int64_t ifull110 = ifull100 + ystride;
+        int64_t ifull111 = ifull110 + zstride;
 
 # ifdef DEBUG
-        long ifullmax = compressed_index->size() - 1;
+        int64_t ifullmax = compressed_index->size() - 1;
         if (
             (ifull000 < 0) || (ifull000 > ifullmax) ||
             (ifull001 < 0) || (ifull001 > ifullmax) ||
@@ -98,17 +98,17 @@ double cal::mpi_atm_sim::interp(double x, double y, double z, std::vector <long>
         }
 # endif // ifdef DEBUG
 
-        long i000 = (*compressed_index)[ifull000];
-        long i001 = (*compressed_index)[ifull001];
-        long i010 = (*compressed_index)[ifull010];
-        long i011 = (*compressed_index)[ifull011];
-        long i100 = (*compressed_index)[ifull100];
-        long i101 = (*compressed_index)[ifull101];
-        long i110 = (*compressed_index)[ifull110];
-        long i111 = (*compressed_index)[ifull111];
+        int64_t i000 = (*compressed_index)[ifull000];
+        int64_t i001 = (*compressed_index)[ifull001];
+        int64_t i010 = (*compressed_index)[ifull010];
+        int64_t i011 = (*compressed_index)[ifull011];
+        int64_t i100 = (*compressed_index)[ifull100];
+        int64_t i101 = (*compressed_index)[ifull101];
+        int64_t i110 = (*compressed_index)[ifull110];
+        int64_t i111 = (*compressed_index)[ifull111];
 
 # ifdef DEBUG
-        long imax = realization->size() - 1;
+        int64_t imax = realization->size() - 1;
         if (
             (i000 < 0) || (i000 > imax) ||
             (i001 < 0) || (i001 > imax) ||
